Added JUnit tests for Getters and Setters

+++ src/main/java/com/staed/ers/beans/Reimbursement.java
@@ -71,7 +71,7 @@ public class Reimbursement {
 		return amount;
 	}
 	
-	public void setAmount(int amount) {
+	public void setAmount(float amount) {
 		this.amount = amount;
 	}
